adds caching to associated_users.
It's typical for any recent OS/UA to activate multiple HTTP connections when a network connections becomes available: each of this connections generates a middleware query that, in turn, generates a query to the openVPN management service and a query to OWUMS.
This caching strategy reduces the overall load of the aforementioned process and mitigates DoS attacks.

Note: another (although more complex) way to achieve this could be caching (page not http caching).